+++ src/main/java/com/example/invc_proj/dto/ReceiptResponseDTO.java
@@ -1,13 +1,15 @@
 package com.example.invc_proj.dto;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 public class ReceiptResponseDTO {
     private int receiptId;
-    private int invoiceId;
-    private long amount;
+    private Long invoiceId;
+    private BigDecimal amount;
     private LocalDateTime paymentDate;
     private String paymentMode;
     private String reference;
