+++ src/main/java/com/example/invc_proj/dto/ServiceCostRequest.java
@@ -9,12 +9,12 @@ import java.math.BigDecimal;
 public class ServiceCostRequest {
         private int serviceId;        // The ID of the service selected by the user
         private String serviceName;    // The name of the service selected by the user
-        private int serviceCost; // The cost of the service entered by the user
-        private int totalCost;  // the total cost of the services entered
+        private BigDecimal serviceCost; // The cost of the service entered by the user
+        private BigDecimal totalCost;  // the total cost of the services entered
 
     // Getters and Setters
 
-         public ServiceCostRequest(int serviceId,int serviceCost,String serviceName,int totalCost)
+         public ServiceCostRequest(int serviceId,BigDecimal serviceCost,String serviceName,BigDecimal totalCost)
          {
              this.serviceId = serviceId;
              this.serviceName = serviceName;
